+++ gd-gltg/webpack.config.dev.js
@@ -1,7 +1,22 @@
 // @flow
+const Webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 
+const commonConfigBase = require('../webpack.config.common');
 const devConfigBase = require('../webpack.config.dev');
-const commonConfigCore = require('./webpack.config.common');
+const commonConfigGLTG = require('./webpack.config.common');
 
-module.exports = webpackMerge(commonConfigCore, devConfigBase);
+module.exports = webpackMerge(
+    commonConfigBase,
+    devConfigBase,
+    commonConfigGLTG,
+    {
+        plugins: [
+            new Webpack.DefinePlugin({
+                'process.env.NODE_ENV': '"development"',
+                'process.env.GEOSERVER_URL': JSON.stringify(process.env.GEOSERVER_URL || 'https://gltg-dev.ncsa.illinois.edu/geoserver'),
+                'process.env.GEOSTREAMS_URL': JSON.stringify(process.env.GEOSTREAMS_URL || 'https://gltg-dev.ncsa.illinois.edu/geostreams')
+            })
+        ]
+    }
+);